Deleting non-existing transients gave error - fixed
# Conflicts:
#	src/Products.php
#	src/Vouchers.php

+++ src/ContactForm.php
@@ -152,7 +152,9 @@ class ContactForm
         foreach ($forms as $id) {
             $errors += self::deleteTransients($subdomain, $id);
         }
-        $errors += $recrasPlugin->transients->delete($subdomain . '_contactforms');
+        if ($recrasPlugin->transients->get($subdomain . '_contactforms')) {
+            $errors += $recrasPlugin->transients->delete($subdomain . '_contactforms');
+        }
 
         return $errors;
     }
@@ -172,8 +174,10 @@ class ContactForm
 
         $errors = 0;
 
-        $errors += $recrasPlugin->transients->delete($subdomain . '_contactform_' . $formID . '_v2');
-        $errors += $recrasPlugin->transients->delete($subdomain . '_contactform_' . $formID . '_fields');
+        $name = $subdomain . '_contactform_' . $formID . '_v2';
+        if ($recrasPlugin->transients->get($name)) {
+            $errors += $recrasPlugin->transients->delete($name);
+        }
 
         return $errors;
     }
